% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_leaf_depth.R
\name{average_leaf_depth}
\alias{average_leaf_depth}
\title{Calculate the average leaf depth statistic. The average leaf depth statistic
is a normalized version of the Sackin index, normalized by the number of
tips.}
\usage{
average_leaf_depth(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
gamma statistic
}
\description{
Calculate the average leaf depth statistic. The average leaf depth statistic
is a normalized version of the Sackin index, normalized by the number of
tips.
}
\examples{
simulated_tree <- ape::rphylo(n = 10, birth = 1, death = 0)
average_leaf_depth(simulated_tree)
}
\references{
K.-T. Shao and R. R. Sokal. Tree balance. Systematic Zoology,
39(3):266, 1990. doi: 10.2307/2992186.
}
