% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_vertex_depth.R
\name{avg_vert_depth}
\alias{avg_vert_depth}
\title{Average vertex depth metric}
\usage{
avg_vert_depth(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Average depth (in number of edges)
}
\description{
The average vertex depth metric, measures the average path
(in edges), between the tips and the root.
}
\references{
C. Colijn and J. Gardy. Phylogenetic tree shapes resolve disease
transmission patterns. Evolution, Medicine, and Public Health,
2014(1):96-108, 2014. ISSN 2050-6201. doi: 10.1093/emph/eou018.
}
