% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{avgLadder}
\alias{avgLadder}
\title{calculate avgLadder index, from the phyloTop package}
\usage{
avgLadder(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
average number of ladders
}
\description{
calculate avgLadder index, from the phyloTop package
}
