% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b1.R
\name{b1}
\alias{b1}
\title{B1 metric}
\usage{
b1(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object or ltable}

\item{normalization}{"none" or "tips", in which case the resulting
statistic is divided by the number of tips in the tree, as a crude way of
normalization.}
}
\value{
B1 statistic
}
\description{
Balance metric (in the case of a binary tree), which measures
the sum across all internal nodes of one over the maximum depth of all
attached tips to that node.
When calculated on a binary tree, the b1 statistic is a true balance
statistic (meaning that the caterpillar tree is the most extreme, unbalanced
tree). However, the treestats package also supports calculation on
non-binary trees.
}
\references{
K.-T. Shao and R. R. Sokal. Tree Balance.
Systematic Zoology, 39(3):266, 1990. doi: 10.2307/2992186.
}
