% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta}
\alias{beta}
\title{fast function using C++ to calculate the aldous beta statistic}
\usage{
beta(phy)
}
\arguments{
\item{phy}{phylogeny or ltable}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate the aldous beta statistic
}
