% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_statistic}
\alias{beta_statistic}
\title{fast function using C++ to calculate the aldous beta statistic}
\usage{
beta_statistic(
  phy,
  upper_lim = 10,
  algorithm = "COBYLA",
  abs_tol = 1e-04,
  rel_tol = 1e-06
)
}
\arguments{
\item{phy}{phylogeny or ltable}

\item{upper_lim}{upper limit for beta parameter}

\item{algorithm}{optimization algorithm used, default is "COBYLA"
(Constrained Optimization BY Linear Approximations), also available are
"subplex" and "simplex".}

\item{abs_tol}{absolute stopping criterion. Default is 1e-4.}

\item{rel_tol}{relative stopping criterion. Default is 1e-6}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate the aldous beta statistic
}
