% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_statistic_raw}
\alias{beta_statistic_raw}
\title{fast function using C++ to calculate the aldous beta statistic}
\usage{
beta_statistic_raw(phy, upper_lim = 10)
}
\arguments{
\item{phy}{phylogeny or ltable}

\item{upper_lim}{upper limit for beta parameter}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate the aldous beta statistic
}
