% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blum.R
\name{blum}
\alias{blum}
\title{fast function using C++ to calculate the blum index of (im)balance}
\usage{
blum(phy)
}
\arguments{
\item{phy}{phylogeny or ltable}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate the blum index of (im)balance
}
\references{
M. G. B. Blum and O. Francois (2006). Which random processes
describe the Tree of Life? A large-scale study of phylogenetic tree
imbalance. Systematic Biology. 55:685-691.
}
