% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{branching_times}
\alias{branching_times}
\title{function to calculate branching times of a tree, approx 20 times faster
than the DDD equivalent.}
\usage{
branching_times(phy)
}
\arguments{
\item{phy}{phylo object}
}
\description{
function to calculate branching times of a tree, approx 20 times faster
than the DDD equivalent.
}
