% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_all_stats.R
\name{calc_all_stats}
\alias{calc_all_stats}
\title{function to apply all available tree statistics to a single tree}
\usage{
calc_all_stats(phylo)
}
\arguments{
\item{phylo}{phylo object}
}
\value{
list with statistics
}
\description{
this function applies all available tree statistics to a
single tree, being:
\itemize{
  \item{gamma}
  \item{sackin}
  \item{colless}
  \item{beta}
  \item{blum}
  \item{crown age}
  \item{tree height}
  \item{pigot's rho}
  \item{mean branch length}
  \item{number of lineages}
  \item{laplacian spectrum}
  \item{nLTT with empty tree}
  \item{phylogenetic diversity}
  \item{avgLadder index}
  \item{cherries}
  \item{ILnumber}
  \item{pitchforks}
  \item{stairs}
}
}
