% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_all_stats_large_tree.R
\name{calc_all_stats_large_tree}
\alias{calc_all_stats_large_tree}
\title{Apply all available tree statistics to a single tree, but exclude those
statistics that require large amounts of memory.}
\usage{
calc_all_stats_large_tree(phylo, normalize = FALSE)
}
\arguments{
\item{phylo}{phylo object}

\item{normalize}{if set to TRUE, results are normalized (if possible) under
either the  Yule expectation (if available), or the number of tips}
}
\value{
List with statistics
}
\description{
this function applies all tree statistics available in
this package to a single tree, excluding statistics that require large
amounts of memory, because they are based on a distance matrix. The
statistics EXCLUDED are:
\itemize{
  \item laplacian spectrum
  \item variance of pairwise distance (vpd)
  \item maximum eigen vector value
  \item minimum eigenvalue of the Laplacian matrix
  \item maximum eigenvalue of the Laplacian matrix
  \item minimum eigenvalue of the adjacency matrix
  \item maximum eigenvalue of the adjacency matrix
}
}
