% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_balance_stats.R
\name{calc_balance_stats}
\alias{calc_balance_stats}
\title{function to apply all balance statistics to a single tree}
\usage{
calc_balance_stats(phylo, normalize = FALSE)
}
\arguments{
\item{phylo}{phylo object}

\item{normalize}{if set to TRUE, results are normalized (if possible) under
either the  Yule expectation (if available), or the number of tips}
}
\value{
list with statistics
}
\description{
this function applies all tree statistics available in
this package to a single tree, being:
\itemize{
  \item{Sackin}
  \item{Colless}
  \item{Aldous' beta statistic}
  \item{Blum}
  \item{Average Ladder Size}
  \item{cherries}
  \item{ILnumber}
  \item{pitchforks}
  \item{stairs}
  \item{stairs2}
  \item{B1}
  \item{B2}
  \item{area per pair (aPP) }
  \item{average leaf depth (aLD)}
  \item{I statistic}
  \item{ewColless}
  \item{max Delta Width (maxDelW)}
  \item{maximum of Depth}
  \item{variance of Depth}
  \item{maximum Width}
  \item{Rogers}
  \item{total Cophenetic distance}
  \item{symmetry Nodes}
  \item{rquartet index}
  \item{j_one statistic}
}
}
