% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_colless_cpp}
\alias{calc_colless_cpp}
\title{cpp version of colless}
\usage{
calc_colless_cpp(phy, normalization)
}
\arguments{
\item{phy}{phy}

\item{normalization}{n}
}
\description{
cpp version of colless
}
