% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpanda.R
\name{calc_lapl_spectrum}
\alias{calc_lapl_spectrum}
\title{function to calculate the laplacian spectrum, from RPANDA}
\usage{
calc_lapl_spectrum(phy)
}
\arguments{
\item{phy}{phy}
}
\value{
list
}
\description{
function to calculate the laplacian spectrum, from RPANDA
}
