% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpanda.R
\name{calc_lapl_spectrum2}
\alias{calc_lapl_spectrum2}
\title{alternative version, using Rcpp::Eigen}
\usage{
calc_lapl_spectrum2(phy)
}
\arguments{
\item{phy}{phy}
}
\value{
list with five components: 1) eigenvalues the vector of eigen
values, 2) principal_eigenvalue the largest eigenvalueof the spectral
density distribution 3) asymmetry the skewness of the spectral density
distribution 4) peak_height the largest y-axis valueof the spectral
density distribution and 5) eigengap theposition ofthe largest
difference between eigenvalues, giving the number of modalities in the tree.
}
\description{
alternative version, using Rcpp::Eigen
}
\references{
Eric Lewitus, Helene Morlon, Characterizing and Comparing
Phylogenies from their Laplacian Spectrum, Systematic Biology, Volume 65,
Issue 3, May 2016, Pages 495–507, https://doi.org/10.1093/sysbio/syv116
}
