% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_nltt_cpp2}
\alias{calc_nltt_cpp2}
\title{function to calculate nltt}
\usage{
calc_nltt_cpp2(phy1, phy2)
}
\arguments{
\item{phy1}{phylogeny}

\item{phy2}{phylogeny}
}
\description{
function to calculate nltt
}
