% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebalance.R
\name{calc_treebalance_stats}
\alias{calc_treebalance_stats}
\title{performs all treebalance measures}
\usage{
calc_treebalance_stats(phylo)
}
\arguments{
\item{phylo}{phylo object}
}
\value{
crown age
}
\description{
This function can be used as a fast-track power user method to
quickly assess all balance measures provided in the treebalance package.
}
