% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_binary}
\alias{check_binary}
\title{Faster than the ape version}
\usage{
check_binary(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
boolean, true if binary
}
\description{
Fast Rcpp function to check if a tree is binary
}
