% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{cherries}
\alias{cherries}
\title{calculate number of cherries, from the phyloTop package}
\usage{
cherries(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
number of cherries
}
\description{
calculate number of cherries, from the phyloTop package
}
