% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colless.R
\name{colless}
\alias{colless}
\title{calculate colless index, using apTreeshape}
\usage{
colless(phy, norm = NULL)
}
\arguments{
\item{phy}{phylo object}

\item{norm}{A character string equals to NULL (default) for no
normalization or one of "pda" or "yule".}
}
\value{
colless index
}
\description{
calculate colless index, using apTreeshape
}
