% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colless.R
\name{colless3}
\alias{colless3}
\title{fast function using C++ to calculate the colless index of (im)balance}
\usage{
colless3(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object}

\item{normalization}{A character string equals to NULL (default) for no
normalization or one of "pda" or "yule".}
}
\value{
colless index
}
\description{
fast function using C++ to calculate the colless index of (im)balance
}
