% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colless.R
\name{colless_branch}
\alias{colless_branch}
\title{Branch weighted Colless index of (im)balance.}
\usage{
colless_branch(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
branch weighted colless index
}
\description{
The Branch Colless index is calculated as the sum of
\eqn{(bl_L - bl_R)} over all nodes, where bl_X is the summed branch length
attached to that daughter node.
}
