% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter.R
\name{diameter}
\alias{diameter}
\title{Fast function using C++ to calculate the diameter}
\usage{
diameter(phy, weight = TRUE)
}
\arguments{
\item{phy}{phylo object or ltable}

\item{weight}{if TRUE, uses branch lengths.}
}
\value{
Diameter
}
\description{
The Diameter of a tree is defined as the maximum length of a
shortest path. When taking branch lenghts into account, this is typically
twice the crown age. When not taking branch lengths into account, this may
be linked to max_depth
}
\references{
Chindelevitch, Leonid, et al. "Network science inspires novel
tree shape statistics." Plos one 16.12 (2021): e0259877.
}
