% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_cherries.R
\name{double_cherries}
\alias{double_cherries}
\title{Double Cherry index}
\usage{
double_cherries(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
number of cherries
}
\description{
Calculate the number of double cherries, where a single cherry
is a node connected to two tips, and a double cherry is a node connected
to two cherry nodes.
}
\references{
Chindelevitch, Leonid, et al. "Network science inspires novel
tree shape statistics." PloS one 16.12 (2021): e0259877.
}
