% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvector.R
\name{eigen_vector}
\alias{eigen_vector}
\title{Fast function using C++ to calculate eigen vector centrality}
\usage{
eigen_vector(phy, weight = TRUE, scale = FALSE)
}
\arguments{
\item{phy}{phylo object or ltable}

\item{weight}{if TRUE, uses branch lengths.}

\item{scale}{if TRUE, the eigenvector is rescaled}
}
\value{
List with the eigen vector and the leading eigen value
}
\description{
Eigen vector centrality associates with each node v the positive
value e(v), such that: \eqn{sum_{e~v} w(uv) * e(u) = \lambda * e(v)}. Thus,
e(v) is the Perron-Frobenius eigenvector of the adjacency matrix of the tree.
}
\references{
Chindelevitch, Leonid, et al. "Network science inspires novel
tree shape statistics." Plos one 16.12 (2021): e0259877.
}
