% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp_data}
\alias{emp_data}
\alias{sim_data}
\title{empirical results}
\format{
list with two entries: cor containing the correlation calculated on
the full dataset and data, containing the data.
}
\usage{
sim_data
}
\description{
Summary statistics for empirical tree dataset
}
\keyword{datasets}
