% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_stat.R
\name{entropy_j}
\alias{entropy_j}
\title{Intensive quadratic entropy statistic J.}
\usage{
entropy_j(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
intensive quadratic entropy statistic J
}
\description{
The intensive quadratic entropy statistic J is given by the
average distance between two randomly chosen species, thus given by the
sum of all pairwise distances, divided by \eqn{S^2}, where S is the number
of tips of the tree.
}
\references{
Izsák, János, and Laszlo Papp. "A link between ecological
diversity indices and measures of biodiversity."
Ecological Modelling 130.1-3 (2000): 151-156.
}
