% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\title{gamma statistic, from ape package}
\usage{
gamma(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
gamma statistic
}
\description{
gamma statistic, from ape package
}
