% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_statistic}
\alias{gamma_statistic}
\title{calculate the gamma statistic, using C++.}
\usage{
gamma_statistic(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
gamma statistic
}
\description{
calculate the gamma statistic, using C++.
}
