% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbal_steps.R
\name{imbalance_steps}
\alias{imbalance_steps}
\title{imbalance steps index}
\usage{
imbalance_steps(input_obj, normalize = FALSE)
}
\arguments{
\item{input_obj}{phylo object or ltable}

\item{normalize}{if true, the number of steps taken is divided by the tree
size}
}
\value{
required number of moves
}
\description{
Calculates the number of moves required to transform the focal
tree into a fully imbalanced (caterpillar) tree. Higher value indicates a
more balanced tree.
}
