% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_ltable}
\alias{is_ltable}
\title{function to determine of object is ltable
this function is kind of crude, as it only checks some aspects that make an
ltable}
\usage{
is_ltable(x)
}
\description{
function to determine of object is ltable
this function is kind of crude, as it only checks some aspects that make an
ltable
}
\keyword{internal}
