% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2phylo.R
\name{l_to_phylo}
\alias{l_to_phylo}
\title{Convert an L table to phylo object}
\usage{
l_to_phylo(ltab, drop_extinct = TRUE)
}
\arguments{
\item{ltab}{ltable}

\item{drop_extinct}{should extinct species be dropped from the phylogeny?}
}
\value{
phylo object
}
\description{
Convert an L table to phylo object
}
