% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2phylo_ed.R
\name{l_to_phylo_ed}
\alias{l_to_phylo_ed}
\title{Convert an reverse-timescale L table to phylo object,
only use it with PDD simulation}
\usage{
l_to_phylo_ed(ltab, t, drop_extinct = TRUE)
}
\arguments{
\item{ltab}{ltable}

\item{t}{simulation time when converting it to phylogeny}

\item{drop_extinct}{should extinct species be dropped from the phylogeny?}
}
\value{
phylo object
}
\description{
Convert an reverse-timescale L table to phylo object,
only use it with PDD simulation
}
