% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_statistics.R
\name{list_statistics}
\alias{list_statistics}
\title{Provides a list of all available statistics in the package}
\usage{
list_statistics(only_balance_stats = FALSE)
}
\arguments{
\item{only_balance_stats}{only return those statistics associated with
measuring balance of a tree}
}
\value{
vector with names of summary statistics
}
\description{
Provides a list of all available statistics in the package
}
