% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2phylo.R
\name{ltable_to_newick}
\alias{ltable_to_newick}
\title{Convert an L table to newick string}
\usage{
ltable_to_newick(ltab, drop_extinct = TRUE)
}
\arguments{
\item{ltab}{ltable}

\item{drop_extinct}{should extinct species be dropped from the phylogeny?}
}
\value{
phylo object
}
\description{
Convert an L table to newick string
}
