% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unbalanced_tree3.R
\name{make_unbalanced_tree3}
\alias{make_unbalanced_tree3}
\title{this function increasingly increases the imbalance of a tree}
\usage{
make_unbalanced_tree3(init_tree, unbal_steps)
}
\arguments{
\item{init_tree}{starting tree to work with}

\item{unbal_steps}{number of imbalance generating steps}
}
\value{
phylo object
}
\description{
the goal of this function is to increasingly imbalance a tree,
by changing the topology, one move at a time. It does so by re-attaching
terminal branches to the root lineage, through the ltable. In effect, this
causes the tree to become increasingly caterpillarlike. When started with
a balanced tree, this allows for exploring the gradient between a fully
balanced tree, and a fully unbalanced tree.
Please note that the algorithm will try to increase imbalance, until a fully
caterpillar like tree is reached, which may occur before unbal_steps is
reached.
}
\examples{
simulated_tree <- ape::rphylo(n = 16, birth = 1, death = 0)
brts <- branching_times(simulated_tree)
balanced_tree <- nodeSub::create_balanced_tree(brts)
unbalanced_tree <- nodeSub::create_unbalanced_tree(brts)
intermediate_tree <- make_unbalanced_tree(balanced_tree, 8)
colless(balanced_tree)
colless(intermediate_tree) # should be intermediate value
colless(unbalanced_tree) # should be highest colless value
}
