% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_betweenness.R
\name{max_betweenness}
\alias{max_betweenness}
\title{Fast function using C++ to calculate maximum betweenness centrality.}
\usage{
max_betweenness(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Maximum Betweenness
}
\description{
Betweenness centrality associates with each node v, the two
nodes v, w, for which the shortest path between v and w runs through v, if
the tree were re-rooted at node v. Then, we report the node with maximum
betweenness centrality.
}
\references{
Chindelevitch, Leonid, et al. "Network science inspires novel
tree shape statistics." Plos one 16.12 (2021): e0259877.
}
