% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_depth.R
\name{max_depth}
\alias{max_depth}
\title{Maximum depth metric}
\usage{
max_depth(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object or ltable}

\item{normalization}{"none" or "tips", in which case the resulting statistic
is divided by the number of tips in the tree.}
}
\value{
Maximum depth (in number of edges)
}
\description{
The maximum depth metric, measures the maximal path (in edges),
between the tips and the root.
}
\references{
C. Colijn and J. Gardy. Phylogenetic tree shapes resolve disease
transmission patterns. Evolution, Medicine, and Public Health,
2014(1):96-108, 2014. ISSN 2050-6201. doi: 10.1093/emph/eou018.
}
