% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ladder.R
\name{max_ladder}
\alias{max_ladder}
\title{Maximum ladder index}
\usage{
max_ladder(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
longest ladder in the tree
}
\description{
Calculate the maximum ladder index, from the phyloTop package.
Higher values indicate more unbalanced trees. To calculate the maximum ladder
index, first all potential ladders in the tree are calculated. A ladder is
defined as a sequence of nodes where one of the daughter branches is a
terminal branch, resulting in a 'ladder' like pattern. The maximum ladder
index then represents the longest ladder found among all observed ladders
in the tree.
}
