% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_width.R
\name{max_width}
\alias{max_width}
\title{Fast function using C++ to calculate the sackin index of (im)balance.}
\usage{
max_width(phy)
}
\arguments{
\item{phy}{phylogeny or ltable}
}
\value{
Sackin index
}
\description{
The Sackin index is calculated as the sum of ancestors for each
of the tips. Higher values indicate higher imbalance. Two normalizations
are available, where a correction is made for tree size, under either a
Yule expectation, or a pda expectation.
}
\references{
M. J. Sackin (1972). "Good" and "Bad" Phenograms.
Systematic Biology. 21:225-226.
}
