% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{mean_branch_length}
\alias{mean_branch_length}
\title{calculate the mean branch length}
\usage{
mean_branch_length(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
mean branch length
}
\description{
calculate the mean branch length
}
