% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{mean_branch_length_ext}
\alias{mean_branch_length_ext}
\title{Mean length of external branch lengths of a tree, e.g. of
branches leading to a tip.}
\usage{
mean_branch_length_ext(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
mean of external branch lengths
}
\description{
Mean length of external branch lengths of a tree, e.g. of
branches leading to a tip.
}
