% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTT.R
\name{nLTT_base}
\alias{nLTT_base}
\title{calculate the nLTT, using a reference 'empty' tree with only two lineages.
Only use if you are very certain
about the input data. If you are less certain, use the nLTT function from
the nLTT package.}
\usage{
nLTT_base(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
number of lineages
}
\description{
calculate the nLTT, using a reference 'empty' tree with only two lineages.
Only use if you are very certain
about the input data. If you are less certain, use the nLTT function from
the nLTT package.
}
