% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_lineages.R
\name{number_of_lineages}
\alias{number_of_lineages}
\title{calculate the number of lineages}
\usage{
number_of_lineages(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
number of lineages
}
\description{
calculate the number of lineages
}
