% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phylo_to_l}
\alias{phylo_to_l}
\title{function to generate ltable from phy object}
\usage{
phylo_to_l(phy)
}
\arguments{
\item{phy}{phylo object}
}
\description{
function to generate ltable from phy object
}
