% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic_diversity.R
\name{phylogenetic_diversity}
\alias{phylogenetic_diversity}
\title{calculate phylogenetic diversity, e.g. the total branch length of the extant
tree}
\usage{
phylogenetic_diversity(phy, t = 1e+10)
}
\arguments{
\item{phy}{phylo object}

\item{t}{time point at which to measure phylogenetic diversity}
}
\value{
phylogenetic diversity
}
\description{
calculate phylogenetic diversity, e.g. the total branch length of the extant
tree
}
