% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{pitchforks}
\alias{pitchforks}
\title{calculate pitchforks, from the phyloTop package}
\usage{
pitchforks(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
number of pitchforks
}
\description{
calculate pitchforks, from the phyloTop package
}
