% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precalc_Laplacian_mem}
\alias{precalc_Laplacian_mem}
\title{precompute laplacian stuff}
\usage{
precalc_Laplacian_mem(edge, w, nNodes)
}
\arguments{
\item{edge}{edge}

\item{w}{weights}

\item{nNodes}{nnodes}
}
\description{
precompute laplacian stuff
}
