% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prep_lapl_spec}
\alias{prep_lapl_spec}
\title{function to create laplacian matrix}
\usage{
prep_lapl_spec(phy)
}
\arguments{
\item{phy}{phy}
}
\value{
numericmatrix
}
\description{
function to create laplacian matrix
}
