% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psv.R
\name{psv}
\alias{psv}
\title{Fast function using C++ to calculate Phylogenetic Species Variability.}
\usage{
psv(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Phylogenetic Species Variability
}
\description{
The phylogenetic species variability is bounded in [0, 1]. The
psv quantifies how phylogenetic relatedness decrease the variance of a
(neutral) trait shared by all species in the tree. As species become more
related, the psv tends to 0.
}
\references{
Helmus M.R., Bland T.J., Williams C.K. & Ives A.R. (2007)
Phylogenetic measures of biodiversity. American Naturalist, 169, E68-E83
}
