% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase_ltable.R
\name{rebase_ltable}
\alias{rebase_ltable}
\title{a function to modify an ltable, such that the longest path in the phylogeny
is a crown lineage.}
\usage{
rebase_ltable(ltable)
}
\arguments{
\item{ltable}{ltable}
}
\value{
modified ltable
}
\description{
a function to modify an ltable, such that the longest path in the phylogeny
is a crown lineage.
}
