% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_imbalance.R
\name{root_imbalance}
\alias{root_imbalance}
\title{Root imbalance}
\usage{
root_imbalance(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Root imbalance
}
\description{
Measures the distribution of tips over the two crown lineages,
e.g. \eqn{n1 / (n1 + n2)}, where n1 is the number of tips connected to crown
lineage 1 and n2 is the number of tips connected to crown lineage 2. We
always take n1 > n2, thus root imbalance is always in [0.5, 1]. Root
imbalance is also calculated for polytomous trees, as long as there is no
polytomy at the root.
}
\references{
Guyer, Craig, and Joseph B. Slowinski. "Adaptive radiation and
the topology of large phylogenies." Evolution 47.1 (1993): 253-263.
}
