% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquartet.R
\name{rquartet}
\alias{rquartet}
\title{Rquartet index.}
\usage{
rquartet(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object or ltable}

\item{normalization}{The index can be normalized by the expectation under
the Yule ("yule") or PDA model ("pda").}
}
\value{
rquartet index
}
\description{
The rquartet index counts the number of potential fully
balanced rooted subtrees of 4 tips in the tree. The function in treestats
assumes a bifurcating tree. For trees with polytomies, we refer the user to
treebalance::rquartedI, which can also take polytomies into account.
}
\references{
T. M. Coronado, A. Mir, F. Rosselló, and G. Valiente. A balance
index for phylogenetic trees based on rooted quartets. Journal of
Mathematical Biology, 79(3):1105-1148, 2019. doi: 10.1007/s00285-019-01377-w.
}
