% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sackin.R
\name{sackin}
\alias{sackin}
\title{fast function using C++ to calculate sackin index, please note that the
sackin index is not normalized here.}
\usage{
sackin(phy, norm = "none")
}
\arguments{
\item{phy}{phylogeny or ltable}

\item{norm}{normalization, either 'none' (default), "yule" or "pda".}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate sackin index, please note that the
sackin index is not normalized here.
}
