% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bd_tree.R
\name{sim_bd_tree}
\alias{sim_bd_tree}
\title{function to simulate a birth-death tree, using a variety of packages}
\usage{
sim_bd_tree(
  birth,
  death,
  num_trees,
  max_t = NULL,
  max_lin = NULL,
  complete = TRUE,
  method = "TreeSim"
)
}
\arguments{
\item{birth}{birth rate}

\item{death}{death rate}

\item{num_trees}{number of trees to simulate}

\item{max_t}{total time}

\item{max_lin}{total number of extant lineages at present}

\item{complete}{if true, returns a tree including extinct lineages}

\item{method}{chosen package used to simulate the tree, available are:
'TreeSim', 'ape', 'TESS', 'DDD', 'geiger' and 'phytools'.}
}
\value{
list of trees
}
\description{
when max_t is set, the tree is simulated conditional on
the maximum time. Similarly, when max_lin is set, the tree is simulated
conditional on the number of extant lineages. Although some packages (e.g.
TreeSim and TESS) provide functions to simulate conditional on both time
and number of lineages, this functionality is not universally available and
not supported here.
}
