% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_density.R
\name{spectral_density}
\alias{spectral_density}
\title{calculate the spectra of eigenvalues for the modified graph Laplacian
of a phylogenetic tree. These are taken from the RPANDA package.}
\usage{
spectral_density(phy, meth = "standard")
}
\arguments{
\item{phy}{phylo object}

\item{meth}{method used to calculate the spectral density, either 'standard'
or 'normal'. Standard calculates the unnormalized spectral density, 'normal'
normalizes the spectral density to the degree matrix.}
}
\value{
a list with the following values:
\itemize{
\item{eigenvalues} the vector of eigenvalues
\item{principal_eigenvalue} the largest eigenvalue
\item{asymmetry} the skewness of the psectral density profile
\item{peak_height} the larges y-axis value of the spectral density profile
\item{eigengap} the position of the largest difference between eigenvalues,
given the number of modalities in the tree.
}
}
\description{
calculate the spectra of eigenvalues for the modified graph Laplacian
of a phylogenetic tree. These are taken from the RPANDA package.
}
\references{
Lewitus, E., Morlon, H., Characterizing and comparing phylogenies
from their Laplacian spectrum, Systematic Biology, Volume 65, Issue 3,
2016, Pages 495–507, https://doi.org/10.1093/sysbio/syv116
}
