% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{stairs}
\alias{stairs}
\title{calculate number of stairs, from the phyloTop package}
\usage{
stairs(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
number of stairs
}
\description{
calculate number of stairs, from the phyloTop package
}
