% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_coph.R
\name{tot_coph}
\alias{tot_coph}
\title{Fast function using C++ to calculate the total cophenetic index.}
\usage{
tot_coph(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object or ltable}

\item{normalization}{"none" or "yule", when "yule" is chosen, the statistic
is divided by the Yule expectation}
}
\value{
Total cophenetic index
}
\description{
The total cophenetic index is the sum of the depth of the last
common ancestor of all pairs of leaves.
}
\references{
A. Mir, F. Rosselló, and L. Rotger. A new balance index for
phylogenetic trees. Mathematical Bio-sciences, 241(1):125-136, 2013.
doi: 10.1016/j.mbs.2012.10.005.
}
