% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_age.R
\name{tree_height}
\alias{tree_height}
\title{Height of a tree.}
\usage{
tree_height(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
crown age
}
\description{
In a reconstructed tree, obtaining the tree height is fairly
straightforward, and the function beautier::get_crown_age does a great job
at it. However, in a non-ultrametric tree, that function no longer works.
Alternatively, taking the maximum value of adephylo::distRoot will also yield
the tree height (including the root branch), but will typically perform
many superfluous calculations and thus be slow.
}
