% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeness.R
\name{treeness}
\alias{treeness}
\title{Treeness statistic}
\usage{
treeness(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
sum of all internal branch lengths (e.g. branches not leading
to a tip) divided by the sum over all branch lengths.
}
\description{
Calculates the fraction of tree length on internal branches,
also known as treeness or stemminess
}
