\name{treestats-package}
\alias{treestats-package}
\alias{treestats}
\docType{package}
\title{
  Collection of phylogenetic tree statistics
}
\description{
  The treestats package contains a collection of phylogenetic tree statistics,
  implemented in C++ to ensure high speed.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Maintainer: Thijs Janzen <thijsjanzen@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}

