% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{var_branch_length}
\alias{var_branch_length}
\title{Variance of branch lengths of a tree,
including extinct branches.}
\usage{
var_branch_length(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
variance of branch lengths
}
\description{
Variance of branch lengths of a tree,
including extinct branches.
}
