% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{var_branch_length_ext}
\alias{var_branch_length_ext}
\title{Variance of external branch lengths of a tree, e.g. of
branches leading to a tip.}
\usage{
var_branch_length_ext(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
variance of external branch lengths
}
\description{
Variance of external branch lengths of a tree, e.g. of
branches leading to a tip.
}
