% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{var_branch_length_int}
\alias{var_branch_length_int}
\title{Calculates the variance of internal branch lengths of a tree, e.g. of
branches not leading to a tip.}
\usage{
var_branch_length_int(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
variance of internal branch lengths
}
\description{
Calculates the variance of internal branch lengths of a tree, e.g. of
branches not leading to a tip.
}
