% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_leaf_depth.R
\name{var_leaf_depth}
\alias{var_leaf_depth}
\title{Fast function using C++ to calculate the variance of leaf depth statistic}
\usage{
var_leaf_depth(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Variance of leaf depths
}
\description{
The variance of leaf depth statistic returns the variance
of depths across all tips.
}
\references{
T. M. Coronado, A. Mir, F. Rosselló, and L. Rotger.
On Sackin's original proposal: the variance of the leaves' depths as a
phylogenetic balance index. BMC Bioinformatics, 21(1), 2020.
doi: 10.1186/s12859-020-3405-1.
}
