% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_pair_dist.R
\name{var_pair_dist}
\alias{var_pair_dist}
\title{Fast function using C++ to calculate the variance of all pairwise distances.}
\usage{
var_pair_dist(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Variance in pairwise distance
}
\description{
After calculating all pairwise distances between all tips,
this function takes the variance across these values.
}
\references{
Webb, C., D. Ackerly, M. McPeek, and M. Donoghue. 2002.
Phylogenies and community ecology. Annual Review of Ecology and
Systematics 33:475-505.
}
