% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLadder.R
\name{avg_ladder}
\alias{avg_ladder}
\title{Average ladder index}
\usage{
avg_ladder(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
average number of ladders
}
\description{
Calculate the avgLadder index, from the phyloTop package.
Higher values indicate more unbalanced trees. To calculate the average ladder
index, first all potential ladders in the tree are calculated. A ladder is
defined as a sequence of nodes where one of the daughter branches is a
terminal branch, resulting in a 'ladder' like pattern. The average ladder
index then represents the average lenght across all observed ladders in the
tree.
}
