% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{avgLadder}
\alias{avgLadder}
\title{Calculate the avgLadder index, from the phyloTop package}
\usage{
avgLadder(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
average number of ladders
}
\description{
Calculate the avgLadder index, from the phyloTop package
}
