% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLadder.R
\name{avgLadder}
\alias{avgLadder}
\title{Calculate the avgLadder index, from the phyloTop package.
Higher values indicate more unbalanced trees.}
\usage{
avgLadder(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
average number of ladders
}
\description{
Calculate the avgLadder index, from the phyloTop package.
Higher values indicate more unbalanced trees.
}
