% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b1.R
\name{b1}
\alias{b1}
\title{Fast function using C++ to calculate the B1 metric}
\usage{
b1(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object or ltable}

\item{normalization}{"none" or "tips", in which case the resulting
statistic is divided by the number of tips in the tree, as a crude way of
normalization.}
}
\value{
Maximum depth (in number of edges)
}
\description{
Balance metric that uses the Shannon-Wiener statistic of
information content. The b2 measure is given by the sum over the depths of
all tips, divided by 2^depth: sum Ni / 2^Ni
}
\references{
K.-T. Shao and R. R. Sokal. Tree Balance.
Systematic Zoology, 39(3):266, 1990. doi: 10.2307/2992186.
}
