% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_times.R
\name{branching_times}
\alias{branching_times}
\title{Branching times of a tree}
\usage{
branching_times(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
vector of branching times
}
\description{
C++ based alternative to
\link[ape:branching.times]{ape::branching.times}.
}
