% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_times.R
\name{branching_times}
\alias{branching_times}
\title{Calculates branching times of a tree, using C++}
\usage{
branching_times(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
vector of branching times
}
\description{
C++ based alternative to `ape::branching.times`, please note
that to maximise speed, `treestats::branching_times` does not return
node names associated to the branching times, in contrast to the ape version.
}
