% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{cherries}
\alias{cherries}
\title{Calculate number of cherries, from the phyloTop package. A cherry is a pair
of sister tips.}
\usage{
cherries(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
number of cherries
}
\description{
Calculate number of cherries, from the phyloTop package. A cherry is a pair
of sister tips.
}
