% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colless.R
\name{colless}
\alias{colless}
\title{fast function using C++ to calculate the colless index of (im)balance}
\usage{
colless(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object}

\item{normalization}{A character string equals to NULL (default) for no
normalization or one of "pda" or "yule".}
}
\value{
colless index
}
\description{
fast function using C++ to calculate the colless index of (im)balance
}
\references{
Colless D H. 1982. Review of: Phylogenetics: The Theory and
Practice of Phylogenetic Systematics. Systematic Zoology 31:100-104.
}
