% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_age.R
\name{crown_age}
\alias{crown_age}
\title{Calculates the crown age of a tree.}
\usage{
crown_age(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
crown age
}
\description{
Calculates the crown age of a tree.
}
