% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_prong.R
\name{four_prong}
\alias{four_prong}
\title{Four prong index}
\usage{
four_prong(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
number of 4-tip caterpillars
}
\description{
Calculate the number of 4-tip caterpillars.
}
\references{
Chindelevitch, Leonid, et al. "Network science inspires novel
tree shape statistics." PloS one 16.12 (2021): e0259877.
Rosenberg, Noah A. "The mean and variance of the numbers of r-pronged nodes
and r-caterpillars in Yule-generated genealogical trees."
Annals of Combinatorics 10 (2006): 129-146.
}
