% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{ILnumber}
\alias{ILnumber}
\title{Calculate ILnumber, from the phyloTop package. The ILnumber is the number
of internal nodes with a single tip child.}
\usage{
ILnumber(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
ILnumber
}
\description{
Calculate ILnumber, from the phyloTop package. The ILnumber is the number
of internal nodes with a single tip child.
}
