% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{ILnumber}
\alias{ILnumber}
\title{calculate ILnumber, from the phyloTop package}
\usage{
ILnumber(phy)
}
\arguments{
\item{phy}{phylo or multiPhylo object}
}
\value{
ILnumber
}
\description{
calculate ILnumber, from the phyloTop package
}
