% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbal_steps.R
\name{imbalance_steps}
\alias{imbalance_steps}
\title{imbalance steps index}
\usage{
imbalance_steps(input_obj, normalization = FALSE)
}
\arguments{
\item{input_obj}{phylo object or ltable}

\item{normalization}{if true, the number of steps taken is normalized by tree
size, by dividing by the maximum number of moves required to move from a
fully balanced to a fully imbalanced tree, which is N - log2(N) - 1, where
N is the number of extant tips.}
}
\value{
required number of moves
}
\description{
Calculates the number of moves required to transform the focal
tree into a fully imbalanced (caterpillar) tree. Higher value indicates a
more balanced tree.
}
