% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_branch_dist.R
\name{inv_branch_dist}
\alias{inv_branch_dist}
\title{Sum of Inverse branch length}
\usage{
inv_branch_dist(phy, add_one = FALSE)
}
\arguments{
\item{phy}{phylo object or ltable}

\item{add_one}{should we take 1 / bl or 1 / (1 + bl) ? if TRUE, then
1 / (1 + bl) is used. Default: FALSE.}
}
\description{
This statistic calculates per tip, the sum of the inverse of
all branch lengths of the shortest path between the root and the tip.
}
\references{
Williams, P. H., Alonso-Alonso, P., Arbetman, M., Françoso,
E., Ghisbain, G., Huang, J., Orr, M. C., Ren, Z.-X., Streinzer, M., T
hanoosing, C., Vandame, R., Waite, M., & Brace, S. (2026). Evolutionary
Tree for All Bumblebee Species World-Wide Estimated by Combining
Information from Fast-Evolving Genes, Slow-Evolving Genes, and Genomic
Data (Apidae, Bombus). Insects, 17(6), 540.
https://doi.org/10.3390/insects17060540
}
