% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_branch_dist.R
\name{inv_branch_dist}
\alias{inv_branch_dist}
\title{Sum of Inverse branch length}
\usage{
inv_branch_dist(phy, add_one = FALSE)
}
\arguments{
\item{phy}{phylo object or ltable}

\item{add_one}{should we take 1 / bl or 1 / (1 + bl) ? if TRUE, then
1 / (1 + bl) is used. Default: TRUE}
}
\value{
Distribution of the sum of inverse brnch length.
}
\description{
This statistic calculates per tip, the sum of the inverse of
all branch lengths of the shortest path between the root and the tip.
}
