% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_branch_dist.R
\name{inv_branch_dist}
\alias{inv_branch_dist}
\title{Sum of Inverse branch length}
\usage{
inv_branch_dist(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Distribution of the sum of inverse brnch length.
}
\description{
This statistic calculates per tip, the sum of the inverse of
all branch lengths of the shortest path between the root and the tip.
}
