% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{mean_branch_length}
\alias{mean_branch_length}
\title{Mean branch length of a tree, including extinct branches.}
\usage{
mean_branch_length(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
mean branch length
}
\description{
Mean branch length of a tree, including extinct branches.
}
