% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_branch_length.R
\name{mean_branch_length_int}
\alias{mean_branch_length_int}
\title{Mean length of internal branches of a tree, e.g. of
branches not leading to a tip.}
\usage{
mean_branch_length_int(phy)
}
\arguments{
\item{phy}{phylo object or Ltable}
}
\value{
mean of internal branch lengths
}
\description{
Mean length of internal branches of a tree, e.g. of
branches not leading to a tip.
}
