% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pair_dist.R
\name{mean_pair_dist}
\alias{mean_pair_dist}
\title{Fast function using C++ to calculate the mean pairwise distance.}
\usage{
mean_pair_dist(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Mean pairwise distance
}
\description{
The mean pairwise distance calculates the average distance
between all combinations of tips.
}
\references{
Webb, C., D. Ackerly, M. McPeek, and M. Donoghue. 2002.
Phylogenies and community ecology. Annual Review of Ecology and
Systematics 33:475-505.
}
