% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_laplace.R
\name{minmax_laplace}
\alias{minmax_laplace}
\title{Laplacian Matrix properties}
\usage{
minmax_laplace(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
List with the minimum and maximum eigenvalues
}
\description{
Calculates the eigenvalues of the Laplacian Matrix, where the
Laplacian matrix is the matrix representation of a graph, in this case a
phylogeny.
}
\references{
Chindelevitch, Leonid, et al. "Network science inspires novel
tree shape statistics." Plos one 16.12 (2021): e0259877.
}
