% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mntd.R
\name{mntd}
\alias{mntd}
\title{Mean Nearest Taxon distance}
\usage{
mntd(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Mean Nearest Taxon Distance.
}
\description{
Per tip, evaluates the shortest distance to another tip,
then takes the average across all tips.
}
\references{
Webb, C., D. Ackerly, M. McPeek, and M. Donoghue. 2002.
Phylogenies and community ecology. Annual Review of Ecology and
Systematics 33:475-505.
}
