% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_over_md.R
\name{mw_over_md}
\alias{mw_over_md}
\title{Maximum width of branch depths divided by the maximum depth}
\usage{
mw_over_md(phy)
}
\arguments{
\item{phy}{phylogeny or ltable}
}
\description{
Calculates the maximum width divided by the maximum depth.
}
\references{
C. Colijn and J. Gardy. Phylogenetic tree shapes resolve disease
transmission patterns. Evolution, Medicine, and Public Health,
2014(1):96-108, 2014. ISSN 2050-6201. doi: 10.1093/emph/eou018.
}
