% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLTT.R
\name{nLTT}
\alias{nLTT}
\title{calculate the nLTT, using C++. Only use if you are very certain
about the input data. If you are less certain, use the nLTT function from
the nLTT package.}
\usage{
nLTT(phy, ref_tree)
}
\arguments{
\item{phy}{phylo or multiPhylo object}

\item{ref_tree}{reference tree to compare with}
}
\value{
number of lineages
}
\description{
calculate the nLTT, using C++. Only use if you are very certain
about the input data. If you are less certain, use the nLTT function from
the nLTT package.
}
