% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_lineages.R
\name{number_of_lineages}
\alias{number_of_lineages}
\title{Number of tips of a tree, including extinct tips.}
\usage{
number_of_lineages(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
number of lineages
}
\description{
Number of tips of a tree, including extinct tips.
}
