% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchforks.R
\name{pitchforks}
\alias{pitchforks}
\title{Calculate pitchforks, from the phyloTop package, a pitchfork is a clade
with three tips.}
\usage{
pitchforks(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
number of pitchforks
}
\description{
Calculate pitchforks, from the phyloTop package, a pitchfork is a clade
with three tips.
}
