% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{pitchforks}
\alias{pitchforks}
\title{Calculate pitchforks, from the phyloTop package, a pitchfork is a clade
with three tips.}
\usage{
pitchforks(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
number of pitchforks
}
\description{
Calculate pitchforks, from the phyloTop package, a pitchfork is a clade
with three tips.
}
