% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchforks.R
\name{pitchforks}
\alias{pitchforks}
\title{Calculate pitchforks, from the phyloTop package, a pitchfork is a clade
with three tips.}
\usage{
pitchforks(input_obj, normalization = "none")
}
\arguments{
\item{input_obj}{phylo object or ltable}

\item{normalization}{"none" or "tips", in which case the found number of
pitchforks is divided by the expected number.}
}
\value{
number of pitchforks
}
\description{
Calculate pitchforks, from the phyloTop package, a pitchfork is a clade
with three tips.
}
