% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchforks.R
\name{pitchforks}
\alias{pitchforks}
\title{Number of pitchforks}
\usage{
pitchforks(input_obj, normalization = "none")
}
\arguments{
\item{input_obj}{phylo object or ltable}

\item{normalization}{"none" or "tips", in which case the found number of
pitchforks is divided by the expected number.}
}
\value{
number of pitchforks
}
\description{
Pitchforks are a clade with three tips, as introduced in the phyloTop
package.
}
