% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sackin.R
\name{sackin}
\alias{sackin}
\title{fast function using C++ to calculate sackin index}
\usage{
sackin(phy)
}
\arguments{
\item{phy}{phylogeny or ltable}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate sackin index
}
