% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sackin.R
\name{sackin}
\alias{sackin}
\title{fast function using C++ to calculate the sackin index of (im)balance}
\usage{
sackin(phy, normalization = "none")
}
\arguments{
\item{phy}{phylogeny or ltable}

\item{normalization}{normalization, either 'none' (default), "yule" or "pda".}
}
\value{
sackin index
}
\description{
fast function using C++ to calculate the sackin index of (im)balance
}
\references{
M. J. Sackin (1972). "Good" and "Bad" Phenograms.
Systematic Biology. 21:225-226.
}
