% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stairs.R
\name{stairs}
\alias{stairs}
\title{Calculates the staircase-ness measure, from the phyloTop package. The
staircase-ness reflects the number of subtrees that are imbalanced, e.g.
subtrees where the left child has more extant tips than the right child, or
vice versa.}
\usage{
stairs(input_obj)
}
\arguments{
\item{input_obj}{phylo object or ltable}
}
\value{
number of stairs
}
\description{
Calculates the staircase-ness measure, from the phyloTop package. The
staircase-ness reflects the number of subtrees that are imbalanced, e.g.
subtrees where the left child has more extant tips than the right child, or
vice versa.
}
\references{
Norström, Melissa M., et al. "Phylotempo: a set of r scripts for
assessing and visualizing temporal clustering in genealogies inferred from
serially sampled viral sequences." Evolutionary Bioinformatics 8 (2012):
EBO-S9738.
}
