% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_internal_path.R
\name{tot_internal_path}
\alias{tot_internal_path}
\title{Total internal path length}
\usage{
tot_internal_path(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Total internal path length
}
\description{
The total internal path length describes the sums of the depths
of all inner vertices of the tree.
}
\references{
Knuth, Donald E. The Art of Computer Programming:
Fundamental Algorithms, volume 1. Addison-Wesley Professional, 1997.
}
