% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_path.R
\name{tot_path_length}
\alias{tot_path_length}
\title{Total  path length}
\usage{
tot_path_length(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Total path length
}
\description{
The total path length describes the sums of the depths
of all vertices of the tree.
}
\references{
C. Colijn and J. Gardy. Phylogenetic tree shapes resolve disease
transmission patterns. Evolution, Medicine, and Public Health,
2014(1):96-108, 2014. ISSN 2050-6201. doi: 10.1093/emph/eou018.
}
