% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_leaf_depth.R
\name{var_leaf_depth}
\alias{var_leaf_depth}
\title{Fast function using C++ to calculate maximum depth metric}
\usage{
var_leaf_depth(phy)
}
\arguments{
\item{phy}{phylo object or ltable}
}
\value{
Maximum depth (in number of edges)
}
\description{
The maximum depth metric, measures the maximal path (in edges),
between the tips and the root.
}
\references{
T. M. Coronado, A. Mir, F. Rosselló, and L. Rotger.
On Sackin's original proposal: the variance of the leaves' depths as a
phylogenetic balance index. BMC Bioinformatics, 21(1), 2020.
doi: 10.1186/s12859-020-3405-1.
}
