% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_leaf_depth.R
\name{var_leaf_depth}
\alias{var_leaf_depth}
\title{Variance of leaf depth statistic}
\usage{
var_leaf_depth(phy, normalization = "none")
}
\arguments{
\item{phy}{phylo object or ltable}

\item{normalization}{"none" or "yule", when "yule" is chosen, the statistic
is divided by the Yule expectation}
}
\value{
Variance of leaf depths
}
\description{
The variance of leaf depth statistic returns the variance
of depths across all tips, where depth of a tip indicates the distance of
the tip to the root.
}
\references{
T. M. Coronado, A. Mir, F. Rosselló, and L. Rotger.
On Sackin's original proposal: the variance of the leaves' depths as a
phylogenetic balance index. BMC Bioinformatics, 21(1), 2020.
doi: 10.1186/s12859-020-3405-1.
}
